集成 gorm

+++ data/db.go
@@ -26,17 +26,17 @@ type DB struct {
 // UserServices 是一个包含账号, 用户和开票相关所有方法的接口
 type UserServices interface {
 	SignUp(email, password string) (*model.Account, error)
-	ChangePassword(id, accountID int64, passwd string) error
-	AddToken(accountID, userID int64, name string) (*model.AccessToken, error)
-	RemoveToken(accountID, userID, tokenID int64) error
-	Auth(accountID int64, token string, pat bool) (*model.Account, *model.User, error)
+	ChangePassword(id, accountID uint, passwd string) error
+	AddToken(accountID, userID uint, name string) (*model.AccessToken, error)
+	RemoveToken(accountID, userID, tokenID uint) error
+	Auth(accountID uint, token string, pat bool) (*model.Account, *model.User, error)
 	GetUserByEmail(email string) (*model.User, error)
-	GetDetail(id int64) (*model.Account, error)
+	GetDetail(id uint) (*model.Account, error)
 	GetByStripe(stripeID string) (*model.Account, error)
-	SetSeats(id int64, seats int) error
-	ConvertToPaid(id int64, stripeID, subID, plan string, yearly bool, seats int) error
-	ChangePlan(id int64, plan string, yearly bool) error
-	Cancel(id int64) error
+	SetSeats(id uint, seats int) error
+	ConvertToPaid(id uint, stripeID, subID, plan string, yearly bool, seats int) error
+	ChangePlan(id uint, plan string, yearly bool) error
+	Cancel(id uint) error
 }
 
 // AdminServices TODO: investigate this...
@@ -46,14 +46,14 @@ type AdminServices interface {
 
 // WebhookServices 一个包含管理 webhook 所有方法的接口
 type WebhookServices interface {
-	Add(accountID int64, events, url string) error
-	List(accountID int64) ([]model.Webhook, error)
-	Delete(accountID int64, event, url string) error
+	Add(accountID uint, events, url string) error
+	List(accountID uint) ([]model.Webhook, error)
+	Delete(accountID uint, event, url string) error
 	AllSubscriptions(event string) ([]model.Webhook, error)
 }
 
 // NewID 根据帐户和用户ID返回每秒的唯一字符串。
-func NewID(accountID, userID int64) string {
+func NewID(accountID, userID uint) string {
 	n := time.Now()
 	i, _ := strconv.Atoi(
 		fmt.Sprintf("%d%d%d%d%d%d%d%d",
